requirements.txt
alembic==1.13.2
altgraph==0.17.4
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bottle==0.12.25
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
pefile==2023.2.7
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PySocks==1.7.1
python-dotenv==1.0.1
pythonnet==3.0.3
pywebview==5.1
pywin32-ctypes==0.2.2
requests==2.32.3
selenium==4.22.0
setuptools==70.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
